sanitize during rendering of markdown
dompurify is the first stage of sanitizing, but it still enables submitting potentially dangerous javascript to the database. removing all parentheses is a sure-fire way to sanitize, but removes a lot of markdown features, e.g. alt-text links. but this (through marked options) is imho a better solution for sanitizing. the post content (i.e. potentially dangerous javascript code) should never be touched anyway